# m3ntorship-Posts-1.0.0-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Post Service API Specs\n",
    "version": "1.0.0",
    "title": "Pickly Post Service"
  },
...
  "security": [
    {
      "firebaseAuth": [

      ]
    }
  ],
  "paths": {
    "/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Create a new post",
        "operationId": "create",
        "responses": {
          "201": {
            "description": "Post Created Successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "isAnonymous": {
                    "type": "boolean"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "get All posts for feed.\n",
        "description": "posts owned by user are different\nfrom ones that aren't,\nposts owned by user will always have their votes.\nnon-owners only get votes when they have already voted.\n\nIn the example, both objects are shown.\n",
        "operationId": "getPosts",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "If-Post-Owner": {
                        "type": "object",
                        "$ref": "#/components/schemas/Post-Owner"
                      },
                      "If-Post-NotOwner": {
                        "type": "object",
                        "$ref": "#/components/schemas/Post-NotOwner"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Post-Owner": {
        "type": "object",
        "required": [
          "caption",
          "resources",
          "isAnonymous"
        ],
        "properties": {
          "caption": {
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/Resource-Owner"
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "ownedByCurrentUser": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Post-NotOwner": {
        "type": "object",
        "required": [
          "caption",
          "resources",
          "isAnonymous"
        ],
        "properties": {
          "caption": {
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/Resource-NotOwner"
          },
          "isAnonymous": {
            "type": "boolean",
            "example": false
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "voted": {
            "type": "boolean"
          },
          "ownedByCurrentUser": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Resource-Owner": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image-Owner"
            }
          }
        }
      },
      "Resource-NotOwner": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image-NotOwner"
            }
          }
        }
      },
      "Image-Owner": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "votes": {
            "$ref": "#/components/schemas/Vote"
          }
        }
      },
      "Image-NotOwner": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "votes": {
            "$ref": "#/components/schemas/Vote"
          },
          "votedByUser": {
            "type": "boolean"
          }
        }
      },
      "Vote": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "default": 0
          }
        }
      },
      "Author": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "firebaseAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "UnAuthorized Access token is missing or invalid"
      }
    }
  },
  "servers": [
    {
      "description": "Try it Now",
      "url": "https://pickly-posts-be-dev.m3ntorship.net/"
    }
  ]
}